src/requirements.txt
sip
pyqt5
pillow
pyinstaller